requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
dill==0.3.4
Django==4.0.3
djangorestframework==3.13.1
greenlet==1.1.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
msgpack==1.0.3
mysqlclient==2.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.13.4
pynvim==0.4.3
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
wrapt==1.14.0
yapf==0.32.0
